requirements.txt
streamlit
pandas
plotly
python-binance
pytz
streamlit-autorefresh
streamlit-option-menu